extracted ComplaintItemColumnInfo component
